removed unused UsersAdministrationController.cs added INhCurrentSessionContext
