+++ shesha-core/src/Shesha.NHibernate/NHibernate/Session/INhCurrentSessionContext.cs
@@ -1,0 +1,16 @@
+﻿using NHibernate;
+using System;
+
+namespace Shesha.NHibernate.Session
+{
+#nullable enable
+    public interface INhCurrentSessionContext : IDisposable
+    {
+        /// <summary>
+        /// 
+        /// </summary>
+        ISession Session { get; }
+        ISession? SessionOrNull { get; }
+    }
+#nullable restore
+}
